+++ lib/src/utils/utils.dart
@@ -1,4 +1,4 @@
-library utils;
+library;
 
 import 'dart:async';
 import 'dart:convert';
@@ -6,7 +6,6 @@ import 'dart:math';
 import 'dart:typed_data';
 import 'dart:developer' as dev;
 
-import 'package:asn1lib/asn1lib.dart';
 import 'package:blockchain_utils/blockchain_utils.dart';
 import 'package:web3dart/crypto.dart';
 import 'package:web3dart/web3dart.dart';