+++ lib/src/web3_signers_base.dart
@@ -1,4 +1,4 @@
-library variance;
+library;
 
 import 'dart:async';
 import 'dart:convert';