+++ lib/web3_signers.dart
@@ -8,4 +8,5 @@ export 'src/interfaces/interfaces.dart'
         SignatureOptions,
         PassKeysOptions;
 export 'src/utils/utils.dart';
+export 'src/vendor/vendor.dart';
 export 'src/web3_signers_base.dart';